MAJOR FORK. Add a new use case document to guide development.
This commit marks a major forking point in the mole code base. The previous
commits were a periodic exploration of using the Todoist Sync API via an
asyncio wrapper to trigger actions based on long-lived stateful syncing
processes.

I am now switching to Motion from Todoist, and hope to restrict the scope of
asyncio in the next codebase as a timewaster and pre-optimization I don't need.
I still do intend to build an execution wrapper that uses an asyncio loop to
dispatch periodic update/sync tasks, but this loop will be kept as high in the
stack as possible and should not "infect" any of the lower level parts of the
API with asynchronicity.