mole 0.6 - Integrates nb-cli Tasks and Todos
The behavior of `mole log`, `mole zonein`, and `mole tasks` now respects
an additional environment variable, `MOLE_TODO`, which can optionally be
set when zoneing-in.

This system will serve as the foundation for an iterative
strategy/tactics/execution/reflection workflow for software development
using mole.

When zoneing-in to a specific todo, an nb-cli .todo.md file will be
created where logging can occur in a context sensitive place. The
project's worklow is still updating with the proper headings and
accessible via nb-cli directly.